[fft-cuda] change the type of `n_cmplx` from double to int

+++ src/fft-hip/main.cu
@@ -129,7 +129,7 @@ int main(int argc, char** argv)
   const int n_ffts = half_n_ffts * 2;
   const int half_n_cmplx = half_n_ffts * 512;
   const unsigned long used_bytes = half_n_cmplx * 2 * sizeof(T2);
-  const double n_cmplx = (double)half_n_cmplx*2.0;
+  const int n_cmplx = half_n_cmplx*2;
 
   fprintf(stdout, "used_bytes=%lu, n_cmplx=%g\n", used_bytes, n_cmplx);
